+++ examples/add-new-command/packages/sdk/src/doc/rpc/sample.ts
@@ -1,0 +1,10 @@
+import type * as common from "./common";
+
+export interface PingRequest extends common.CommandRequest<"ping"> {
+    message?: string;
+}
+
+export interface PingResponse extends common.CommandResponse {
+    data: string;
+    timestamp: string;
+}