+++ packages/sdk/src/doc/rpc/cmds.ts
@@ -1,0 +1,58 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+import type * as common from "./common";
+
+export interface IssueConsoleRequest extends common.CommandRequest<"consoleCommand"> {
+    /**
+     * Console command to execute
+     */
+    commandText: string;
+    /**
+     * Name of the console
+     */
+    consoleName?: string;
+}
+
+export interface IssueConsoleResponse extends common.CommandResponse {
+    /**
+     * Data returned from the console command
+     */
+    data: string;
+}
+
+export interface IssueTsoRequest extends common.CommandRequest<"tsoCommand"> {
+    /**
+     * TSO command to execute
+     */
+    commandText: string;
+}
+
+export interface IssueTsoResponse extends common.CommandResponse {
+    /**
+     * Data returned from the TSO command
+     */
+    data: string;
+}
+
+export interface IssueUnixRequest extends common.CommandRequest<"unixCommand"> {
+    /**
+     * UNIX command to execute
+     */
+    commandText: string;
+}
+
+export interface IssueUnixResponse extends common.CommandResponse {
+    /**
+     * Data returned from the UNIX command
+     */
+    data: string;
+}