+++ packages/sdk/src/doc/types.ts
@@ -9,7 +9,7 @@
  *
  */
 
-import type { CommandResponse } from "./gen/common";
+import type { CommandResponse } from "./rpc/common";
 
 export type B64String = string & { __brand: "B64String" };
 export namespace B64String {